remove unused code in user-features
